+++ app/routes/$time.tsx
@@ -5,7 +5,6 @@ import { StatsCardList } from "~/components/StatsCardList"
 import { TimeSelectBar } from "~/components/TimeSelectBar"
 import { getPeriodSummary } from "~/logic/cost-balancer"
 import { convertMonthStrTo2CharStr, getTimeSelectFormProps } from "~/utils/form"
-import { ErrorBoundary } from "./_index"
 
 export async function loader({ params }: LoaderArgs) {
   const { time } = params
@@ -61,5 +60,3 @@ export default function PeriodSummaryRoute() {
     </main>
   )
 }
-
-ErrorBoundary()
