+++ remix.config.js
@@ -13,10 +13,13 @@ module.exports = {
   // assetsBuildDirectory: "public/build",
   // serverBuildPath: "api/index.js",
   // publicPath: "/build/",
+  tailwind: true,
   future: {
     v2_dev: true,
     v2_routeConvention: true,
     v2_meta: true,
     v2_errorBoundary: true,
+    v2_normalizeFormMethod: true,
+    v2_headers: true,
   },
 }
